refactor: rename lib/api to lib/registry
docker and npm
