+++ lib/config/presets.js
@@ -1,7 +1,7 @@
 const configParser = require('./index');
 const massage = require('./massage');
 const migration = require('./migration');
-const npm = require('../api/npm');
+const npm = require('../registry/npm');
 
 module.exports = {
   resolveConfigPresets,
