+++ lib/registry/docker.js
