+++ lib/registry/npm.js
