+++ lib/workers/package-file/index.js
@@ -1,6 +1,6 @@
 const configParser = require('../../config');
 const depTypeWorker = require('../dep-type');
-const npmApi = require('../../api/npm');
+const npmApi = require('../../registry/npm');
 
 let logger = require('../../logger');
