+++ lib/workers/package/docker.js
@@ -1,5 +1,5 @@
 const semver = require('semver');
-const dockerApi = require('../../api/docker');
+const dockerApi = require('../../registry/docker');
 const versions = require('./versions');
 const compareVersions = require('compare-versions');
