+++ lib/workers/package/npm.js
@@ -1,4 +1,4 @@
-const npmApi = require('../../api/npm');
+const npmApi = require('../../registry/npm');
 const versions = require('./versions');
 
 module.exports = {
