+++ test/workers/package/npm.spec.js
@@ -1,10 +1,10 @@
-const npmApi = require('../../../lib/api/npm');
+const npmApi = require('../../../lib/registry/npm');
 const versions = require('../../../lib/workers/package/versions');
 const npm = require('../../../lib/workers/package/npm');
 const defaultConfig = require('../../../lib/config/defaults').getConfig();
 const logger = require('../../_fixtures/logger');
 
-jest.mock('../../../lib/api/npm');
+jest.mock('../../../lib/registry/npm');
 npmApi.getDependency = jest.fn();
 
 describe('lib/workers/package/npm', () => {
